cheri: Apply the getpage.cocci semantic patch
Calls to get_page variants that return an address instead of
a pointer but be treated as a uintptr_t not as unsigned long.
If the return value is assigned to a variable change its type
to uintptr_t.

Apply like this:

	make COCCI=scripts/coccinelle/cheri/getpage.cocci \
		MODE=patch coccicheck

+++ arch/arm64/kernel/vdso.c
@@ -182,7 +182,7 @@ static int aarch32_alloc_kuser_vdso_page(void)
 {
 	extern char __kuser_helper_start[], __kuser_helper_end[];
 	int kuser_sz = __kuser_helper_end - __kuser_helper_start;
-	unsigned long vdso_page;
+	uintptr_t vdso_page;
 
 	if (!IS_ENABLED(CONFIG_KUSER_HELPERS))
 		return 0;
