+++ arch/parisc/kernel/pci-dma.c
@@ -402,7 +402,7 @@ void *arch_dma_alloc(struct device *dev, size_t size,
 		dma_addr_t *dma_handle, gfp_t gfp, unsigned long attrs)
 {
 	unsigned long vaddr;
-	unsigned long paddr;
+	uintptr_t paddr;
 	int order;
 
 	if (boot_cpu_data.cpu_type != pcxl2 && boot_cpu_data.cpu_type != pcxl)
