+++ drivers/net/ethernet/mediatek/mtk_eth_soc.c
@@ -1260,7 +1260,7 @@ static bool mtk_rx_get_desc(struct mtk_eth *eth, struct mtk_rx_dma_v2 *rxd,
 static void *mtk_max_lro_buf_alloc(gfp_t gfp_mask)
 {
 	unsigned int size = mtk_max_frag_size(MTK_MAX_LRO_RX_LENGTH);
-	unsigned long data;
+	uintptr_t data;
 
 	data = __get_free_pages(gfp_mask | __GFP_COMP | __GFP_NOWARN,
 				get_order(size));
