+++ drivers/net/ethernet/sun/niu.c
@@ -10000,7 +10000,7 @@ static void *niu_phys_alloc_coherent(struct device *dev, size_t size,
 				     u64 *dma_addr, gfp_t flag)
 {
 	unsigned long order = get_order(size);
-	unsigned long page = __get_free_pages(flag, order);
+	uintptr_t page = __get_free_pages(flag, order);
 
 	if (page == 0UL)
 		return NULL;
